Removing unused method

+++ lib/generators/react_particles/install/react/src_generator.rb
@@ -15,28 +15,19 @@ module ReactParticles
 
           def copy_src
             src_template_dir_path = "../src_templates"
-              case self.behavior
-              when :invoke
-                directory src_template_dir_path, src_dir_path
-              when :revoke
-                `rm -rf #{src_dir_path}`
-                puts indent_str("removed ".red) + "#{src_dir_path}/*"
-              end
+            case self.behavior
+            when :invoke
+              directory src_template_dir_path, src_dir_path
+            when :revoke
+              `rm -rf #{src_dir_path}` if (Dir.exists? src_dir_path)
+              puts indent_str("removed ".red) + "#{src_dir_path}/*"
+            end
           end
 
           private
 
             def src_dir_path
               src_dir_path = "app/javascript/#{namespace}/src"
-            end
-
-            def append_to_gitignore(file)
-              if Rails.root.join(".gitignore").exist?
-                append_to_file(".gitignore", "\n #{file} \n")
-              else
-                system "touch .gitignore"
-                append_to_file(".gitignore", "\n #{file} \n")
-              end
             end
 
             def namespace
